ARM: dts: msm: removing "config-recovery" and "drv-name" for qps615
Config recovery should be taken care by individual EP so, this
property is not needed for qps615 switch. Removing this property
from pcie0 instance for RB5 platform. For non-switch usecase this
feature is needed so, enabling it for HDK platform.

DRV feature is not supported for qps615 switch so, removing this
property for RB5 target from both pcie0 & pcie1 instance. This
feature is supported for non-switch usecase so, enabling it for
HDK platform.

Change-Id: Iecf785722b067bedd124dfece67e3ce14e979a7d